# backend/openapi.json
{
  "components": {
    "schemas": {
      "ActivityOut": {
        "properties": {
          "activity_type": {
            "title": "Activity Type",
            "type": "string"
          },
          "concept_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Concept Ids",
            "type": "array"
          },
          "lesson_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lesson Id"
          },
          "presentation_format": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Presentation Format"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "score": {
            "title": "Score",
            "type": "number"
          }
        },
        "required": [
          "activity_type",
          "concept_ids",
          "lesson_id",
          "reason",
          "score"
        ],
        "title": "ActivityOut",
        "type": "object"
      },
...
      "AssessmentOutcomeIn": {
        "properties": {
          "baseline_question_attempt_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Question Attempt Id"
          },
          "is_transfer_question": {
            "default": false,
            "title": "Is Transfer Question",
            "type": "boolean"
          },
          "question_attempt_id": {
            "format": "uuid",
            "title": "Question Attempt Id",
            "type": "string"
          }
        },
        "required": [
          "question_attempt_id"
        ],
        "title": "AssessmentOutcomeIn",
        "type": "object"
      },
...
      "EngagementOutcomeIn": {
        "properties": {
          "extra": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extra"
          },
          "outcome_type": {
            "pattern": "^(VIEWED|COMPLETED|ABANDONED|REJECTED)$",
            "title": "Outcome Type",
            "type": "string"
          }
        },
        "required": [
          "outcome_type"
        ],
        "title": "EngagementOutcomeIn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HelpfulnessFeedbackIn": {
        "properties": {
          "rating": {
            "maximum": 1.0,
            "minimum": -1.0,
            "title": "Rating",
            "type": "integer"
          }
        },
        "required": [
          "rating"
        ],
        "title": "HelpfulnessFeedbackIn",
        "type": "object"
      },
...
      "PresentationOutcomeIn": {
        "properties": {
          "format": {
            "title": "Format",
            "type": "string"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "format",
          "success"
        ],
        "title": "PresentationOutcomeIn",
        "type": "object"
      },
      "PresentationSwitchIn": {
        "properties": {
          "from_format": {
            "title": "From Format",
            "type": "string"
          },
          "to_format": {
            "title": "To Format",
            "type": "string"
          }
        },
        "required": [
          "from_format",
          "to_format"
        ],
        "title": "PresentationSwitchIn",
        "type": "object"
      },
...
      "RecommendationOut": {
        "properties": {
          "alternatives": {
            "items": {
              "$ref": "#/components/schemas/ActivityOut"
            },
            "title": "Alternatives",
            "type": "array"
          },
          "decision_id": {
            "format": "uuid",
            "title": "Decision Id",
            "type": "string"
          },
          "recommended": {
            "$ref": "#/components/schemas/ActivityOut"
          }
        },
        "required": [
          "decision_id",
          "recommended",
          "alternatives"
        ],
        "title": "RecommendationOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Backend Service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/courses/{course_id}/adaptation-decisions/{decision_id}/outcomes/assessment": {
      "post": {
        "description": "The pedagogical-effect path: mastery_delta is computed server-side\n(current mastery minus the value already frozen in the decision's own\ninput_snapshot), never taken from the client. Set is_transfer_question\nwhen the attempt is on a *different* question testing the same concept\nin a new context, not a repeat of the original question.",
        "operationId": "record_assessment_outcome_api_v1_courses__course_id__adaptation_decisions__decision_id__outcomes_assessment_post",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Course Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "decision_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Decision Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-email",
            "required": true,
            "schema": {
              "title": "X-User-Email",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": true,
            "schema": {
              "title": "X-Internal-Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssessmentOutcomeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Assessment Outcome",
        "tags": [
          "adaptation"
        ]
      }
    },
    "/api/v1/courses/{course_id}/adaptation-decisions/{decision_id}/outcomes/engagement": {
      "post": {
        "description": "Plain activity completion/abandonment/viewing, or an explicit\nrejection (the learner picked a different alternative). Never carries a\nmastery_delta or transfer_success -- those only come from\n/outcomes/assessment, which requires real QuestionAttempt evidence.",
        "operationId": "record_engagement_outcome_api_v1_courses__course_id__adaptation_decisions__decision_id__outcomes_engagement_post",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Course Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "decision_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Decision Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-email",
            "required": true,
            "schema": {
              "title": "X-User-Email",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": true,
            "schema": {
              "title": "X-Internal-Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EngagementOutcomeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Engagement Outcome",
        "tags": [
          "adaptation"
        ]
      }
    },
    "/api/v1/courses/{course_id}/adaptation-decisions/{decision_id}/outcomes/helpfulness": {
      "post": {
        "description": "Explicit learner feedback -- self-reported, kept in its own\nsignal_category, never averaged with measured mastery change.",
        "operationId": "record_helpfulness_outcome_api_v1_courses__course_id__adaptation_decisions__decision_id__outcomes_helpfulness_post",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Course Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "decision_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Decision Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-email",
            "required": true,
            "schema": {
              "title": "X-User-Email",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": true,
            "schema": {
              "title": "X-Internal-Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HelpfulnessFeedbackIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Helpfulness Outcome",
        "tags": [
          "adaptation"
        ]
      }
    },
    "/api/v1/courses/{course_id}/adaptation-history": {
      "get": {
        "description": "What was recommended, why, and what happened afterward -- in\nchronological order. Ownership-checked the same way every other\ncourse-scoped resource is: a course the caller doesn't own is 404, not\na filtered-down empty list.",
        "operationId": "get_adaptation_history_api_v1_courses__course_id__adaptation_history_get",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Course Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-email",
            "required": true,
            "schema": {
              "title": "X-User-Email",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": true,
            "schema": {
              "title": "X-Internal-Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Adaptation History",
        "tags": [
          "adaptation"
        ]
      }
    },
...
    "/api/v1/courses/{course_id}/next-activity": {
      "get": {
        "description": "Every call persists an AdaptationDecision before this returns anything\n-- if that write fails, the caller gets a 500, never a silently\nunlogged recommendation (mandate section D).",
        "operationId": "get_next_activity_api_v1_courses__course_id__next_activity_get",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Course Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-email",
            "required": true,
            "schema": {
              "title": "X-User-Email",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": true,
            "schema": {
              "title": "X-Internal-Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Next Activity",
        "tags": [
          "adaptation"
        ]
      }
    },
...
    "/api/v1/presentation-affinity/outcome": {
      "post": {
        "description": "The next checkpoint outcome after viewing a block in `format`.",
        "operationId": "record_presentation_outcome_api_v1_presentation_affinity_outcome_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-user-email",
            "required": true,
            "schema": {
              "title": "X-User-Email",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": true,
            "schema": {
              "title": "X-Internal-Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresentationOutcomeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Presentation Outcome",
        "tags": [
          "adaptation"
        ]
      }
    },
    "/api/v1/presentation-affinity/reset": {
      "post": {
        "description": "Clears only presentation-format preference evidence -- ConceptMastery/\nMasteryEvent rows (what the learner has actually demonstrated knowing)\nare untouched. A learner resetting \"which format works for me\" and a\nlearner erasing \"what I've proven I know\" are different requests with\ndifferent reasons to exist; conflating them would strand real progress\nbehind a preference-reset button.",
        "operationId": "reset_presentation_affinity_api_v1_presentation_affinity_reset_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-user-email",
            "required": true,
            "schema": {
              "title": "X-User-Email",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": true,
            "schema": {
              "title": "X-Internal-Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reset Presentation Affinity",
        "tags": [
          "adaptation"
        ]
      }
    },
    "/api/v1/presentation-affinity/switch": {
      "post": {
        "description": "A learner's manual format override -- weaker evidence than an\noutcome, but never ignored, and never treated as a contradiction to\ncorrect (guardrail).",
        "operationId": "record_presentation_switch_api_v1_presentation_affinity_switch_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-user-email",
            "required": true,
            "schema": {
              "title": "X-User-Email",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": true,
            "schema": {
              "title": "X-Internal-Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresentationSwitchIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Presentation Switch",
        "tags": [
          "adaptation"
        ]
      }
    },
...
  }
}